add error handler

+++ src/professor/professor.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { DeleteResult, Repository } from 'typeorm';
 import { CreateProfessorDTO, UpdateProfessorDTO } from './DTO/professor.dto';
@@ -16,7 +16,32 @@ export class ProfessorService {
     return this.professorRepository.find();
   }
 
-  async createProfessor(professor: CreateProfessorDTO) {
+  async create(professor: CreateProfessorDTO) {
+    if (!professor.name || !professor.password)
+      return new HttpException(
+        'Name and password are required',
+        HttpStatus.CONFLICT,
+      );
+
+    const professorFound = await this.professorRepository.findOne({
+      where: {
+        name: professor.name,
+      },
+    });
+
+    if (professorFound) {
+      return new HttpException(
+        'This username already exists',
+        HttpStatus.CONFLICT,
+      );
+    }
+
+    if (!professor.name || !professor.password)
+      return new HttpException(
+        'Name and password are required',
+        HttpStatus.CONFLICT,
+      );
+
     // Define the expiration period
     const date = new Date();
     date.setDate(date.getDate() + 30);
@@ -34,11 +59,33 @@ export class ProfessorService {
     return this.professorRepository.save(newProf);
   }
 
-  deleteProfessor(id: number) {
-    return this.professorRepository.delete({ id });
+  async delete(id: number) {
+    const professorFound = await this.professorRepository.findOne({
+      where: {
+        id,
+      },
+    });
+
+    if (!professorFound) {
+      return new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    const result = await this.professorRepository.delete({ id });
+    if (result.affected) return { message: 'Delete successfully' };
   }
 
-  updateProfessor(id: number, professor: UpdateProfessorDTO) {
-    return this.professorRepository.update({ id }, professor);
+  async update(id: number, professor: UpdateProfessorDTO) {
+    const professorFound = await this.professorRepository.findOne({
+      where: {
+        id,
+      },
+    });
+
+    if (!professorFound) {
+      return new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+    const result = await this.professorRepository.update({ id }, professor);
+
+    if (result.affected) return { message: 'Update successfully' };
   }
 }
